requirements.txt
torch
scikit-learn
scipy
opencv-python
numpy
matplotlib
cmapy
streamlit
streamlit-image-coordinates
st-pages
torchsort
joblib
Pillow
scikit-image
umap-learn
pandas
tqdm
numba
phate
trimap